+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # changelog
 
+## v7.3.0
+
+- add `add_self_auth_provider_v0`, `add_google_auth_provider_v0`, `unlink_auth_provider_v0`.
+- update `send_verification_email_v0` and `send_reset_password_email_v0` to accept optional `redirect_url`.
+- dependencies
+    - update `square_authentication_helper>=4.3.0`.
+
 ## v7.2.1
 
 - fix uv lock file